requirements.txt
tabula-py==2.8.0
camelot-py[cv]==0.10.1
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
numpy==1.24.3
pytest==7.4.0
pytest-mock==3.10.0